requirements.txt
dbt-bigquery
playwright
pandas
python-dotenv
google-cloud-storage